Decouple BrazeEpic from the RRCP component
This moves the BrazeEpic away from "borrowing" the ContributionsEpic
component to instead use components from @guardian/braze-components,
just like the banner. This is required because we want to start using
some new components to promote newsletters in this space. It was also
fragile to rely on a component from another team/system.

Note that these changes depend on some as yet unreleased changes to
@guardian/braze-components where we've split the top level factory
component BrazeMessageComponent into two - BrazeBannerComponent and
BrazeEndOfArticle component.

Still TODO: pass through a real implementation of subscribeToNewsletter
and also countryCode to the component from braze-components. We should
add countryCode to the common end-of-article component props in that
project. Also add ophanComponentId as a required prop in braze-
components for all end-of-article components and their canRender
functions.